Don't throw exception when running `closeServer`
The method `getUniqueClassesDirFor` can throw `NoSuchFileException` when
the root of the parent classes directory doesn't exist. This happens in
our test suite where `closeServer` typically runs after our test has
removed the temporary projects that it creates under `/tmp` and
therefore it clutters the console log in the CI.

This commit makes sure that if such exception happens, we just ignore it
and carry on with our lives.