Use Scala's execution context as scheduler
We have always used a fork join pool with a limited number of threads
because there was no blocking happening inside the compiler. However,
that's no longer the case with the new version of build pipelining where
we can block in `blockUntilMacroClasspathIsReady`.

We now default on the Scala's execution context that should protect us
against that because it respects blocking contexts and spawns new
continuation threads. I hope this change doesn't bring performance down
so we'll follow up with a fresh run of our benchmarks suite.